+++ src/project/lib/structs/repository_info/commands/add_files.rs
@@ -4,10 +4,15 @@ use git2::IndexAddOption;
 use std::path::Path;
 
 impl RepositoryInfo {
-    pub fn add_files(&self, files: &[&str]) -> Result<()> {
+    pub fn add_files<P>(&self, files: &[P]) -> Result<()>
+    where P: AsRef<Path>
+    {
+        let files = files.iter().map(AsRef::as_ref).collect::<Vec<_>>();
         let mut index = self.repo.index()?;
 
-        if files.len() == 1 && files[0] == "*" {
+        if files.len() == 1
+            && files[0].to_str().unwrap_or("") == "*" 
+        {
             index.add_all(["*"].iter(), IndexAddOption::DEFAULT, None)?
         } else {
             for file in files {