+++ src/project/lib/structs/repository_info/commands/remove_files.rs
@@ -2,11 +2,14 @@ use crate::RepositoryInfo;
 use std::path::Path;
 
 impl RepositoryInfo {
-    pub fn remove_files(&self, files: &[&str]) -> crate::utils::helpers::Result<()> {
+    pub fn remove_files<P>(&self, files: &[P]) -> crate::utils::helpers::Result<()>
+    where
+        P: AsRef<Path>,
+    {
         let mut index = self.repo.index()?;
 
         for file in files {
-            index.remove_path(Path::new(file))?;
+            index.remove_path(Path::new(file.as_ref()))?;
         }
 
         index.write()?;